+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Yet unannounced in Chess Simp Discord
 
+* Aug 29, 2023: New challenge: Chess Simp Discord — "Alphabetical"
 * Aug 26, 2023: New challenge: Chess Simp — "Alternate Between Moving Pawns & Pieces"
 * Aug 20, 2023: New challenge: Chess Simp — "That Was Frustrating"